python_backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mangum==0.17.0

numpy==1.26.4
pydantic==2.6.4
python-multipart==0.0.9

tensorflow-cpu==2.15.0
h5py==3.10.0

librosa==0.10.2.post1
scipy==1.11.4
numba==0.59.1
llvmlite==0.42.0
soundfile==0.12.1
scikit-learn==1.4.2
soxr==0.3.7

